Electron index a& ipc registration refactor
